rename githubservice class to oscargit regenerate api resolvers update github graphql schema replace oscargit requests with typesafe axios requests
